requirements.txt
mcp>=1.0.0
snowflake-connector-python>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
asyncio-mqtt>=0.16.0
PyYAML>=6.0.0
